Test add source

+++ src/AppBundle/Tests/Model/Source.php
@@ -34,22 +34,55 @@ class SourceTest extends KernelTestCase
         $entity = $this->em->getRepository('AppBundle:FeedSource')->findOneBy(['source' => $rssSource]);
         $this->assertNotNull($entity);
 
-        $this->em->remove($entity);
-        $this->em->flush();
+        $this->removeSourceEntity($rssSource);
     }
 
-    public function testAddFacebook()
+    public function testFailAddFacebook()
     {
         $source = new Source();
         $source->setEm($this->em);
-        $facebookSource = 'https://graph.facebook.com/fg.gov.ua/posts?access_token=' . App::FB_ACCESS_TOKEN;
-        $this->assertTrue($source->add($facebookSource, Source::SOURCE_TYPE_FACEBOOK));
+        $facebookSource = 'https://graph.facebook.com/fg.gov.ua/posts?access_token=' . Source::FB_ACCESS_TOKEN;
+        $this->removeSourceEntity($facebookSource);
+        $this->assertFalse($source->add($facebookSource, Source::SOURCE_TYPE_FACEBOOK));
 
         $entity = $this->em->getRepository('AppBundle:FeedSource')->findOneBy(['source' => $facebookSource]);
-        $this->assertNotNull($entity);
+        $this->assertNull($entity);
+    }
 
-        $this->em->remove($entity);
-        $this->em->flush();
+    public function testSuccessAddFacebook()
+    {
+        $source = new Source();
+        $source->setEm($this->em);
+        $facebookSource = 'fg.gov.ua';
+        $this->assertTrue($source->add($facebookSource, Source::SOURCE_TYPE_FACEBOOK));
+
+        $fullSource = 'https://graph.facebook.com/' . $facebookSource . '/posts?access_token=' . Source::FB_ACCESS_TOKEN;
+
+        /** @var SourceEntity $entity */
+        $entity = $this->em->getRepository('AppBundle:FeedSource')->findOneBy(['source' => $fullSource]);
+
+        $partOfFullSource = 'https://graph.facebook.com/fg.gov.ua/posts?access_token=';
+        $result = !is_null($entity) && strpos($entity->getSource(), $partOfFullSource) !== false;
+
+        if (!is_null($entity)) {
+            $this->em->remove($entity);
+            $this->em->flush();
+        }
+
+        $this->assertTrue($result, $entity->getSource() . ' <===> ' . $partOfFullSource);
+    }
+
+    /**
+     * @param $source
+     */
+    private function removeSourceEntity($source)
+    {
+        $entity = $this->em->getRepository('AppBundle:FeedSource')->findOneBy(['source' => $source]);
+
+        if (!is_null($entity)) {
+            $this->em->remove($entity);
+            $this->em->flush();
+        }
     }
